Group version utility functions
