+++ ee/tuf/library_manager_test.go
@@ -594,55 +594,3 @@ func Test_sortedVersionsInLibrary_devBuilds(t *testing.T) {
 	require.Equal(t, middleVersion, validVersions[1], "not sorted")
 	require.Equal(t, newerVersion, validVersions[2], "not sorted")
 }
-
-func Test_versionFromTarget(t *testing.T) {
-	t.Parallel()
-
-	testVersions := []struct {
-		target          string
-		binary          autoupdatableBinary
-		operatingSystem string
-		version         string
-	}{
-		{
-			target:          "launcher/darwin/launcher-0.10.1.tar.gz",
-			binary:          binaryLauncher,
-			operatingSystem: "darwin",
-			version:         "0.10.1",
-		},
-		{
-			target:          "launcher/windows/launcher-1.13.5.tar.gz",
-			binary:          binaryLauncher,
-			operatingSystem: "windows",
-			version:         "1.13.5",
-		},
-		{
-			target:          "launcher/linux/launcher-0.13.5-40-gefdc582.tar.gz",
-			binary:          binaryLauncher,
-			operatingSystem: "linux",
-			version:         "0.13.5-40-gefdc582",
-		},
-		{
-			target:          "osqueryd/darwin/osqueryd-5.8.1.tar.gz",
-			binary:          binaryOsqueryd,
-			operatingSystem: "darwin",
-			version:         "5.8.1",
-		},
-		{
-			target:          "osqueryd/windows/osqueryd-0.8.1.tar.gz",
-			binary:          binaryOsqueryd,
-			operatingSystem: "windows",
-			version:         "0.8.1",
-		},
-		{
-			target:          "osqueryd/linux/osqueryd-5.8.2.tar.gz",
-			binary:          binaryOsqueryd,
-			operatingSystem: "linux",
-			version:         "5.8.2",
-		},
-	}
-
-	for _, testVersion := range testVersions {
-		require.Equal(t, testVersion.version, versionFromTarget(testVersion.binary, filepath.Base(testVersion.target)))
-	}
-}
